requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib[bcrypt]
python-dotenv
qrcode
pillow
pyzbar